+++ CHANGELOG.md
@@ -1,15 +1,38 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
+All notable changes to the cookie banner will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
- 
-## Unreleased
-
-### Added
+## [1.2.0] - Unreleased
 
 ### Changed
 
-### Removed
+- Add Path attribute to cookie strings (Issue [#233](https://github.com/app-illinois/Design-Resources/issues/233))
+- Update Cookie Banner to support Content Security Policy Headers (Issue [#241](https://github.com/app-illinois/Design-Resources/issues/241))
+- Add data-cookie-fetch flag to Cookie Banner to support CSP headers (Issue [#242](https://github.com/app-illinois/Design-Resources/issues/242))
+
+## [1.1.2] - 2025-08-11
+
+### Changed
+
+- The Escape Key will now only change focus if the Cookie Banner is open. (Issue [#230](https://github.com/app-illinois/Design-Resources/issues/230))
+
+## [1.1.1] - 2025-07-31
+
+### Changed
+
+- Sets the cookie at the base domain. Issue (Issue [#226](https://github.com/app-illinois/Design-Resources/issues/226)) & (Issue [#213](https://github.com/app-illinois/Design-Resources/issues/213))
+
+## [1.1.0] - 2025-07-03
+
+### Changed
+
+- Update for compliance accessible keyboard control and color contrast issues. (Issue [#168](https://github.com/app-illinois/Design-Resources/issues/168))
+
+## [1.0.0] - 2025-06-30
+
+### Changed
+
+- Initial release